XANMOD: mm: Raise max_map_count default value

+++ include/linux/mm.h
@@ -187,17 +187,18 @@ static inline void __mm_zero_struct_page(struct page *page)
  *
  * When a program's coredump is generated as ELF format, a section is created
  * per a vma. In ELF, the number of sections is represented in unsigned short.
- * This means the number of sections should be smaller than 65535 at coredump.
+ * This means the number of sections should be smaller than 2147483647 at
+ * coredump.
  * Because the kernel adds some informative sections to a image of program at
  * generating coredump, we need some margin. The number of extra sections is
  * 1-3 now and depends on arch. We use "5" as safe margin, here.
  *
- * ELF extended numbering allows more than 65535 sections, so 16-bit bound is
- * not a hard limit any more. Although some userspace tools can be surprised by
- * that.
+ * ELF extended numbering allows more than 2147483647 sections, so 16-bit bound
+ * is not a hard limit any more. Although some userspace tools can be surprised
+ * by that.
  */
 #define MAPCOUNT_ELF_CORE_MARGIN	(5)
-#define DEFAULT_MAX_MAP_COUNT	(USHRT_MAX - MAPCOUNT_ELF_CORE_MARGIN)
+#define DEFAULT_MAX_MAP_COUNT	(INT_MAX - MAPCOUNT_ELF_CORE_MARGIN)
 
 extern int sysctl_max_map_count;
